small fix/refactoring
